requirements.txt
tqdm
torch
torchvision
pyyaml